requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
openai>=1.40.0
anthropic>=0.42.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
